MB-19576 - forestdb can optionally skip batch transaction
When the forestdb store has been configured to "skip_batch" of true,
then the batch API no longer builds up an actual batch in a
transaction as its approach, but instead performs direct, immediate
forestdb Set()/Delete()'s and uses commit() at the end of the
ExecuteBatch().

This allows higher-level layers (such as moss) to use this feature,
along with snapshots, in order to avoid extra memory overhead when the
batches get very large with forestdb.